feat: enhance user model with wallet and task tracking fields
